requirement.txt
flask
speechrecognition
pyttsx3
python-dotenv
groq
gunicorn
